catalog_retriever/requirements.txt
fastapi
uvicorn
openai
pydantic
langchain
langchain-core
langchain-milvus
pymilvus
numpy
pandas
requests
Pillow
